requirements.txt
beautifulsoup4==4.11.1
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
charset-normalizer==2.1.1
dnspython==2.2.1
idna==3.4
pymongo==4.3.2
python-decouple==3.6
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.12
wincertstore==0.2
xmltodict==0.13.0
